+++ primal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog for `primal`
 
-## 0.1.0.0
+## 0.1.1
+
+* Addition of `offToCount`, `offForType`, `offForProxyTypeOf`, `countToOff`,
+  `countForType` and `countForProxyTypeOf`
+* Deprecation of `offAsProxy` and `countAsProxy`
+
+
+## 0.1.0
 
 * Initial release